Fix stuff after PR
* Naming: styling -> uiConfig
* Moved non type default value away from types-file
